Auth completed

+++ src/libs/helpers/file-processor.ts
@@ -1,10 +1,12 @@
 import * as fs from "fs";
 
 export  const deleteFile = async (fileName: string)  => {
-    await fs.unlink(`${fileName}`, (err) => {
-        if (err) {
-            console.error(err);
-            return err;
-        }
-    });
+    if (fileName){
+        await fs.unlink(`${fileName}`, (err) => {
+            if (err) {
+                console.error(err);
+                return err;
+            }
+        });
+    }
 }
